requirements.txt
flask
flask_socketio
torch
transformers
kafka-python
gunicorn
gevent
